Add soft-delete for hardware models
Instead of deleting the entire row, default to just marking the row
as deleted. This will be important if we want to perform cleanup w/
the workers for deleted hardware items.

If the hardware row is deleted, the workers will still be cascade deleted.
This should help the DB stay in a relatively sane state if we ever want
to clean up deleted hardware records.