Add basic fields for device hw type
This also adds a 'worker_overrides' field on the hardware type,
allowing hardware types to declare force overrides on any worker field
for any worker they have enabled. This allows a hardware type to
provide some constraints on user input. Mostly useful b/c the Blazar
device plugin is kind of generic but only some values make sense when
in combination with other things (e.g., a k8s device MUST be enrolled
in Blazar w/ device_type="k8s", it would not make sense otherwise.)

This is currently NOT exposed to the user in the API validation layer
that this is happening, which is not good. But, the DB at least keeps
changes out with this changeset.