implement hardware delete
Deleted hardware records are soft-deleted, which gives us the opportunity
to preform "true" deletion of the backing resources via the workers.

This works by additionally setting worker tasks to pending upon resource
delete. Each worker will need to process this case individually; this is
currently handled for the Blazar and Balena workers.