switch from poetry to setuptools
This was necessary as poetry doesn't support "upper-constraints".
This causes dependencies to differ between local dev and prod, because
kolla uses `pip install` to install the service.

Some changes to the pyproject.toml were needed for this transition,
mostly aguments to setuptools so that config and data files are included
as before we relied on "magic" behavior from poetry.